fix: play screen game fetch api fixed

+++ lib/core/model/journey_models/journey_level_model.dart
@@ -1,7 +1,7 @@
 class JourneyLevel {
   int start;
   int end;
-  int breakpoint;
+  double breakpoint;
   int pageEnd;
   int level;
